Moved size-related files to the size subdirectory
