+++ src/grandorgue/GOFrame.cpp
@@ -37,6 +37,7 @@
 #include "loader/cache/GOCacheCleaner.h"
 #include "midi/GOMidi.h"
 #include "midi/GOMidiEvent.h"
+#include "size/GOLogicalRect.h"
 #include "sound/GOSound.h"
 #include "temperaments/GOTemperament.h"
 #include "threading/GOMutexLocker.h"
@@ -45,7 +46,6 @@
 #include "GOApp.h"
 #include "GODocument.h"
 #include "GOEvent.h"
-#include "GOLogicalRect.h"
 #include "GOOrgan.h"
 #include "GOOrganController.h"
 #include "GOPath.h"
