+++ src/grandorgue/GOMainWindowData.h
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2023 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -10,7 +10,8 @@
 
 #include <wx/gdicmn.h>
 
-#include "GOLogicalRect.h"
+#include "size/GOLogicalRect.h"
+
 #include "GOSaveableObject.h"
 
 class GOOrganController;
