+++ src/grandorgue/config/GOConfig.h
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2023 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -25,9 +25,9 @@
 #include "settings/GOSettingNumber.h"
 #include "settings/GOSettingStore.h"
 #include "settings/GOSettingString.h"
+#include "size/GOLogicalRect.h"
 #include "temperaments/GOTemperamentList.h"
 
-#include "GOLogicalRect.h"
 #include "GOMidiDeviceConfigList.h"
 #include "GOOrganList.h"
 #include "GOPortsConfig.h"
