+++ src/grandorgue/gui/GOGUIPanel.h
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2023 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -14,8 +14,7 @@
 #include "ptrvector.h"
 
 #include "primitives/GOBitmap.h"
-
-#include "GOGUISizeKeeper.h"
+#include "size/GOGUISizeKeeper.h"
 
 class GOGUIControl;
 class GOGUIDisplayMetrics;
