+++ src/grandorgue/size/GOLogicalRect.h
