+++ src/grandorgue/size/GOResizable.h
